+++ src/application/compiler/helpers/hooks/compiler_hooks.h
@@ -198,8 +198,9 @@
 #define def_set_sprite_pattern (def_set_sprite_color + 3)
 #define def_get_sprite_color (def_set_sprite_pattern + 3)
 #define def_get_sprite_pattern (def_get_sprite_color + 3)
+#define def_set_tile_pattern_buffer (def_get_sprite_pattern + 3)
 
-#define def_usr0 (def_get_sprite_pattern + 3)
+#define def_usr0 (def_set_tile_pattern_buffer + 3)
 #define def_usr1 (def_usr0 + 3)
 #define def_usr2 (def_usr1 + 3)
 #define def_usr2_play (def_usr2 + 3)